Replit Main.cpp Copy Updated
Added a few more options to the Switch Case, these are the Reset Array, Print Array, Merge Sort, and Selection sort. Adjusted how the arrays are being printed, instead of one long list, I've cut it in half and printed both halves next to each other. Added a function to snapshot times, this will be used to measure run time of the function being called. Use case switch for merge and selection sort as an example on how to capture it and print out. AGAIN, this is a Replit version of this .cpp file, the corresponding files with the master .cpp files from Visual Studio aren't needed if ran on Replit. Errors on VS appear for merge sort, hence why I am using Replit. Any issues shoot me a text Raf!